kymotracker: split sampling tests into own file
